Split up sub-protocols
