+++ src/contentScript/commandPopover/useCommandRegistry.ts
@@ -18,7 +18,8 @@
 import { useSyncExternalStore } from "use-sync-external-store/shim";
 import { useCallback } from "react";
 import type CommandRegistry from "@/contentScript/commandPopover/CommandRegistry";
-import type { TextCommand } from "@/platform/platformProtocol";
+
+import { TextCommand } from "@/platform/platformTypes/commandPopoverProtocol";
 
 /**
  * React hook to sync React with the text command registry.
